Update lily58.keymap
Updated CAPS to a toggle macro. CAPS wasn't working as a toggle button (needed &kt not &km), but as it is set with a layer modifier key there wasn't a already set function. Testing the TCAPS macro for toggle functionality in addition to the layer selection.

Updated null keys on layer one to be "&none" instead of "&trans" to keep from mistakenly adding in additional characters when navigating. May move all of the media control buttons to the left side instead of duplicating arrow keys as I only use the right ones.